refactor back to RainViewer service
