+++ test/unit/server/services/procedures/rain-viewer.spec.js
@@ -1,0 +1,8 @@
+const utils = require('../../../../assets/utils');
+
+describe(utils.suiteName(__filename), function() {
+    utils.verifyRPCInterfaces('Hangman', [
+        ['getRadarTimeOffsets'],
+        ['getRadarMap', ['latitude', 'longitude', 'width', 'height', 'zoom', 'timeOffset', 'options']],
+    ]);
+});
